backend/requirements.txt
greenlet==3.1.1
Fastapi == 0.116.1
uvicorn == 0.35.0
alembic == 1.17.2
sqlalchemy == 2.0.43
pydantic-settings == 2.12.0
bcrypt == 4.0.1
passlib[bcrypt] == 1.7.4
python-jose == 3.5.0
asyncpg == 0.30.0
